moves Dao index to pages
